+++ packages/accounts-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Populate `.options.entropySource` for new `InternalAccounts` before publishing ([#5841](https://github.com/MetaMask/core/pull/5841))
+
 ## [29.0.0]
 
 ### Changed